+++ app/api/questions.py
@@ -6,7 +6,8 @@ from typing import Annotated
 from ..config import get_question_service, auth_user_jwt
 from ..exceptions.entity_not_found import EntityNotFoundError
 from ..managers.question_manager import QuestionManager
-from ..schemas.question import Question, QuestionCreate, QuestionAll, Source, Difficulty, Status, Tag
+from ..schemas.pagination import Pagination
+from ..schemas.question import Question, QuestionCreate, Source, Difficulty, Status, Tag
 
 router = APIRouter(
     prefix="/questions",
@@ -25,8 +26,7 @@ async def get_all_questions(
     order: Annotated[str | None, Query()] = None,
     page: Annotated[int | None, Query()] = None,
     per_page: Annotated[int | None, Query()] = None,
-    paginate: Annotated[bool | None, Query()] = None,
-) -> QuestionAll:
+) -> Pagination:
     try:
         return await question_service.get_all_questions(
             source=source,
@@ -38,7 +38,6 @@ async def get_all_questions(
             order=order,
             page=page,
             per_page=per_page,
-            paginate=paginate
         )
     except ValueError as e:
         raise HTTPException(status_code=fastapi.status.HTTP_400_BAD_REQUEST, detail=str(e))