+++ js/components/home/question_panel.jsx
@@ -12,6 +12,8 @@ export default function QuestionPanel() {
   const [difficulties, setDifficulties] = useState([]);
   const [statuses, setStatuses] = useState([]);
   const [tags, setTags] = useState([]);
+  const [sortOptions, setSortOptions] = useState([]);
+  const [orderOptions, setOrderOptions] = useState([]);
 
 
   useEffect(async () => {
@@ -28,7 +30,9 @@ export default function QuestionPanel() {
       sources: true,
       difficulties: true,
       statuses: true,
-      tags: true
+      tags: true,
+      sort_by: true,
+      order: true,
     };
 
     fetch(`/api/metadata?${formatQueries(metadataQuery)}`, req)
@@ -38,6 +42,8 @@ export default function QuestionPanel() {
         setDifficulties(data.difficulties);
         setStatuses(data.statuses);
         setTags(data.tags);
+        setSortOptions(data.sort_by);
+        setOrderOptions(data.order);
       })
       .catch(err => {
         throw err;
@@ -58,14 +64,24 @@ export default function QuestionPanel() {
   };
 
   const getQuestions = async (queries = {}) => {
+    const tagsQueryStr = (tags) => {
+      if (!Array.isArray(tags)) return ""
+      const individualQuery = tags.map(tag => `tags=${tag}`);
+      return individualQuery.join("&");
+    };
+
     const req = {
       method: "GET",
       headers: getReqHeader(),
     }
 
-    return await fetch(`/api/questions?${formatQueries(queries)}`, req)
+    const t = queries.tags;
+    delete queries.tags;
+    let queryStr = formatQueries(queries) + "&";
+    queryStr += tagsQueryStr(t);
+
+    return await fetch(`/api/questions?${queryStr}`, req)
       .then(res => res.json())
-      .then(json => json.data)
       .catch(err => {
         throw err;
       })
@@ -73,7 +89,8 @@ export default function QuestionPanel() {
 
   return (
     <div>
-      <FilterBar sources={sources} difficulties={difficulties} statuses={statuses} tags={tags} searchQuestions={searchQuestions} />
+      <FilterBar sources={sources} difficulties={difficulties} statuses={statuses} tags={tags}
+        sortOptions={sortOptions} orderOptions={orderOptions} searchQuestions={searchQuestions} />
       <QuestionList questions={questions} />
       <PaginationBoxes page={page} totalPages={totalPages} fetchForPage={fetchForPage} />
     </div>