+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## Version 0.0.1
+- Improve query using firebase database utility functions (#42)
+- Add sort_by and ordering to filter in front end (#42)
 - Add a single toast context for children to consume (#41)
 - Add UI for deleting and updating solutions (#40)
 - Display solutions (#39)